requirements.txt
# ───── Core ─────
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1

# ───── LangChain / OpenAI ─────
openai==1.14.2
langchain==0.1.17
langchain-core==0.1.48
langchain-community==0.0.36
langchain-openai==0.0.6
langsmith==0.1.21

# ───── LLM Stack ─────
transformers==4.31.0
tokenizers==0.13.3
huggingface_hub==0.19.4
faiss-cpu==1.11.0
sentence-transformers==2.2.2
accelerate==0.27.2
torch==2.2.2
sentencepiece==0.1.99

# ───── Utilities ─────
pydantic==1.10.13
PyYAML>=5.3
bcrypt==4.1.2
pypdf==3.17.1
protobuf==4.25.3
streamlit==1.32.2
google-auth==2.29.0

google-cloud-firestore==2.20.2
python-multipart